+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ## 0.2.0 (unreleased)
 
-- Added support for Discord as a destination.
+- Added support for Discord and Mattermost as destinations.
 
 - Added `Dockerfile` and Compose file.